+++ src/genetic/packer.cpp
@@ -7,7 +7,7 @@
 #include <algorithm>
 
 Packer::Packer(
-    const std::vector<std::pair<int, int>>& rectangles,
+    const std::vector<std::pair<size_t, size_t>>& rectangles,
     size_t tape_width,
     std::unique_ptr<IParentSelector>&& parent_selection_operator,
     std::unique_ptr<ISelector>&& selection_operator,
